fountain: error upon receiving inconsistent part
`receive` now returns an error when trying to receive a
`Part` whose meta-data is inconsistent with previous ones.

This better separates the above from other use-cases that return
`Ok(false)`, for example when receiving a (consistent) `Part` when
the full message has already been decoded.

Internally, this commit also inlines the initalization portion
of the previous `validate`, such that the latter can now accept
a non-mutating `&self`.